refactored code to follow best practices
