+++ cli_tools/__init__.py
@@ -1,0 +1,2 @@
+from .cowsay_utils import cowsayCLI
+from .figlet_utils import figletCLI
