+++ cli_tools/cowsay_utils.py
@@ -1,0 +1,7 @@
+import cowsay
+
+def cowsayCLI(message):
+    outer = "```"
+    c = cowsay.get_output_string('cow', message)
+    return outer + c + outer
+
