+++ cli_tools/figlet_utils.py
@@ -1,0 +1,5 @@
+import pyfiglet
+
+def figletCLI(message):
+    f = pyfiglet.figlet_format(message)
+    return "```" +  f + "```"
