+++ requirements.txt
@@ -20,5 +20,4 @@ pycparser==2.22
 pyfiglet==1.0.2
 setuptools==70.1.0
 sympy==1.12.1
-wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.3/libexec/wheel-0.43.0-py3-none-any.whl#sha256=80a49545e5e5ef2abeb79ec254447300d41dfb76914fc65efa6da8dd0a766052
 yarl==1.9.4
